Fix damage wraparound if very high damage (#11872)

+++ src/tool.h
@@ -106,11 +106,11 @@ DigParams getDigParams(const ItemGroupList &groups,
 
 struct HitParams
 {
-	s16 hp;
+	s32 hp;
 	// Caused wear
 	u32 wear; // u32 because wear could be 65536 (single-use weapon)
 
-	HitParams(s16 hp_ = 0, u32 wear_ = 0):
+	HitParams(s32 hp_ = 0, u32 wear_ = 0):
 		hp(hp_),
 		wear(wear_)
 	{}
